moved imagetyp_instr to it's own file because it is so large. It will eventually be moved to each individual instrument and the file will be dissolved. started converting getProgInfo.php to python. This is a very large file and requires MySQL to complete, but I will probably make a simple web api when I run into it.
